+++ src/hook/timer/use-cascade-timer.ts
@@ -1,5 +1,7 @@
 import { useCallback, useEffect, useMemo } from 'react';
 import { useRecoilValue, useRecoilCallback } from 'recoil';
+import endedAudioPath from '../../audio/ended.mp3';
+import tickTackAudioPath from '../../audio/ticktack.mp3';
 import { SoundableCascadeTimer } from '../../lib/timer/soundable-cascade-timer';
 import { TimerController } from '../../lib/timer/timer-controller';
 import { statusState, lapRemainState, lapIndexState, offsetState, lapDurationsState } from '../../recoil/cascade-timer';
@@ -15,6 +17,14 @@ export type Options = {
   soundOffset?: number;
   controller?: TimerController;
 };
+
+function useAudio(path: string) {
+  const audio = useMemo(() => new Audio(path), [path]);
+  const play = useCallback(async () => {
+    await audio.play();
+  }, [audio]);
+  return { play };
+}
 
 export function useCascadeTimer(options: Options): UseCascadeTimerResult {
   const lapDurations = useRecoilValue(lapDurationsState);
@@ -63,6 +73,18 @@ export function useCascadeTimer(options: Options): UseCascadeTimerResult {
     const unsubscribe = timer.addEventListener('remainChange', syncStateWithTimer);
     return unsubscribe;
   }, [lapDurations, syncStateWithTimer, timer]);
+
+  // audio の再生
+  const { play: playTickTack } = useAudio(tickTackAudioPath);
+  const { play: playEndedAudio } = useAudio(endedAudioPath);
+  useEffect(() => {
+    const unsubscribe1 = timer.addEventListener('ticktack', () => void playTickTack());
+    const unsubscribe2 = timer.addEventListener('ticktackEnded', () => void playEndedAudio());
+    return () => {
+      unsubscribe1();
+      unsubscribe2();
+    };
+  }, [lapDurations, playEndedAudio, playTickTack, syncStateWithTimer, timer]);
 
   return { start, reset, setOffset: setOffsetForOuter };
 }
